frontend/requirements.txt
# requirements.txt
streamlit==1.28.0
requests==2.31.0
pygit2==1.13.2
python-dotenv==1.0.0
pandas==1.5.3
